Re-organize "tutorial" into "tutorials" to make way for a second and possibly other tutorials.
