+++ docs/tutorials/index.rst
@@ -1,0 +1,26 @@
+=========
+Tutorials
+=========
+
+It's better to describe ``wired`` by showing you some applications that use
+it. Below are tutorials which gradually build up a custom application,
+built to be decoupled, extensible, configurable, and pluggable.
+
+By writing them bit-by-bit, we see the kinds of problems that ``wired`` is
+built to solve, as well as the way to solve those problems.
+
+`Tour of Wired <./tour/index.html>`_
+====================================
+
+This tutorial shows a store with Greeters who greet Customers. It starts from
+a very small beginning. Then, as the application's ambitions increase, more of
+``wired`` is put to use.
+
+
+Tutorial Steps
+==============
+
+.. toctree::
+    :hidden:
+
+    tour/index
