+++ tests/tutorials/tour/test_datastore.py
@@ -5,7 +5,7 @@ from wired import ServiceRegistry
 
 @pytest.fixture
 def settings():
-    from requests_views import Settings
+    from tour.datastore import Settings
 
     settings = Settings(punctuation='!!')
     return settings
@@ -13,14 +13,14 @@ def settings():
 
 @pytest.fixture
 def registry(settings):
-    from requests_views import app_bootstrap
+    from tour.datastore import app_bootstrap
 
     r: ServiceRegistry = app_bootstrap(settings)
     return r
 
 
 def test_sample_interactions(registry):
-    from requests_views import sample_interactions
+    from tour.datastore import sample_interactions
 
     greetings = sample_interactions(registry)
     assert 'Hello Mary !!' == greetings[0]
