+++ tests/tutorials/tour/test_decoupled.py
@@ -5,7 +5,7 @@ from wired import ServiceRegistry
 
 @pytest.fixture
 def settings():
-    from context.app import Settings
+    from tour.decoupled import Settings
 
     settings = Settings(punctuation='!!')
     return settings
@@ -13,7 +13,7 @@ def settings():
 
 @pytest.fixture
 def registry(settings):
-    from context.app import setup
+    from tour.decoupled import setup
 
     r: ServiceRegistry = setup(settings)
     return r
@@ -21,27 +21,27 @@ def registry(settings):
 
 @pytest.fixture
 def default_customer():
-    from context.app import Customer
+    from tour.decoupled import Customer
 
     return Customer(name='Mary')
 
 
 @pytest.fixture
 def french_customer():
-    from context.app import FrenchCustomer
+    from tour.decoupled.custom import FrenchCustomer
 
     return FrenchCustomer(name='Henri')
 
 
 def test_greet_customer(registry, default_customer):
-    from context.app import greet_customer
+    from tour.decoupled import greet_customer
 
     actual = greet_customer(registry, default_customer)
     assert 'Hello Mary !!' == actual
 
 
 def test_greet_french_customer(registry, french_customer):
-    from context.app import greet_customer
+    from tour.decoupled import greet_customer
 
     actual = greet_customer(registry, french_customer)
     assert 'Bonjour Henri !!' == actual
