+++ tests/tutorials/tour/test_overrides.py
@@ -5,7 +5,7 @@ from wired import ServiceRegistry
 
 @pytest.fixture
 def settings():
-    from datastore import Settings
+    from tour.overrides import Settings
 
     settings = Settings(punctuation='!!')
     return settings
@@ -13,15 +13,15 @@ def settings():
 
 @pytest.fixture
 def registry(settings):
-    from datastore import app_bootstrap
+    from tour.overrides import app_bootstrap
 
     r: ServiceRegistry = app_bootstrap(settings)
     return r
 
 
 def test_sample_interactions(registry):
-    from datastore import sample_interactions
+    from tour.overrides import sample_interactions
 
     greetings = sample_interactions(registry)
-    assert 'Hello Mary !!' == greetings[0]
+    assert 'Override Mary !!' == greetings[0]
     assert 'Bonjour Henri !!' == greetings[1]
