+++ tests/tutorials/tour/test_requests_views.py
@@ -5,7 +5,7 @@ from wired import ServiceRegistry
 
 @pytest.fixture
 def settings():
-    from overrides import Settings
+    from tour.requests_views import Settings
 
     settings = Settings(punctuation='!!')
     return settings
@@ -13,15 +13,15 @@ def settings():
 
 @pytest.fixture
 def registry(settings):
-    from overrides import app_bootstrap
+    from tour.requests_views import app_bootstrap
 
     r: ServiceRegistry = app_bootstrap(settings)
     return r
 
 
 def test_sample_interactions(registry):
-    from overrides import sample_interactions
+    from tour.requests_views import sample_interactions
 
     greetings = sample_interactions(registry)
-    assert 'Override Mary !!' == greetings[0]
+    assert 'Hello Mary !!' == greetings[0]
     assert 'Bonjour Henri !!' == greetings[1]
